+++ .prettierrc
@@ -13,6 +13,8 @@
     "",
     "@/router",
     "",
+    "@/contexts",
+    "",
     "@/providers",
     "",
     "@/hooks",
