+++ src/components/theme-changer/ThemeChanger.tsx
@@ -1,9 +1,10 @@
 // src/components/ThemeToggle.tsx
 
-import { ThemeContext } from './themeContext';
 import { MoonIcon, SunIcon } from '@heroicons/react/24/outline';
 import { useContext } from 'react';
 import type { ThemeContextType } from '@/interfaces';
+
+import { ThemeContext } from '@/contexts';
 
 export const ThemeChanger = () => {
   const { theme, toggleTheme } = useContext(ThemeContext) as ThemeContextType;
